sns-trigger-app/requirements.txt
mysql-connector-python
pydantic
pytz